refactor: organiza arquivos de Staff e Animal em pastas dedicadas dentro de pages
